serving/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
numpy==2.0.2
google-cloud-storage==2.19.0
google-cloud-logging==3.5.0
httpx==0.27.0
tensorflow==2.18.0